+++ NEWS.md
@@ -1,5 +1,7 @@
 # waldo (development version)
 
+* Read-only S7 properties are ignored (#219).
+
 # waldo 0.6.1
 
 * Only use special bit64 comparison if package is installed.